ADC implementation

+++ solderingStation.X/hcm.h
@@ -5,12 +5,11 @@
 #define _XTAL_FREQ  20000000
 
 
-
 void systemInit(void);
 
 void adcInit(void);
 
-uint16_t readADC(uint8_t);
+int readADC(uint8_t);
 
 #endif
